pyproject.toml
[tool.poetry]
name = "deposit-project"
version = "0.1.0"
description = ""
authors = ["Isopreene <andko2013@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
Django = "^5.1.1"
python-dateutil = "^2.9.0.post0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^5.0.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
